Fix app debug mode and environments variable

+++ services/security/main.py
@@ -1,3 +1,5 @@
+import os
+
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 from services.security.config.database import Base, engine
@@ -16,6 +18,7 @@ from services.security.models.user_has_roles import UserHasRoles
 #SEEDERS
 from services.security.seeders.seed import seed
 
+debug = os.getenv("DEBUG", "False").lower() == "true"
 app = FastAPI()
 
 origins = [
@@ -29,9 +32,9 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-
 # CREATE TABLES
-Base.metadata.drop_all(bind=engine)
+if debug:
+    Base.metadata.drop_all(bind=engine)
 Base.metadata.create_all(bind=engine)
 # SEEDING
 seed()
